Separated models into packages. Corrected client company and user, but javadoc documentation not finished
