+++ src/org/greenpole/entity/model/user/Unit.java
@@ -1,0 +1,87 @@
+package org.greenpole.entity.model.user;
+
+import java.util.HashSet;
+import java.util.Set;
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlElementWrapper;
+import javax.xml.bind.annotation.XmlRootElement;
+import javax.xml.bind.annotation.XmlSeeAlso;
+import javax.xml.bind.annotation.XmlType;
+import org.greenpole.entity.model.clientcompany.ClientCompany;
+
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+
+/**
+ *
+ * @author Yusuf Samsudeen Babashola (Algorithm)
+ */
+@XmlRootElement
+@XmlAccessorType(XmlAccessType.NONE)
+@XmlType(propOrder = {"unitID","name","users"})
+@XmlSeeAlso(User.class)
+public class Unit {
+    @XmlElement
+    private int unitID;
+    @XmlElement
+    private String name;
+    @XmlElementWrapper(name = "users")
+    private Set users = new HashSet(0);
+    
+    public Unit() {}
+    
+    public Unit(int unitID, String name) {
+        this.unitID = unitID;
+        this.name = name;
+    }
+
+    /**
+     * @return the unitID
+     */
+    public int getUnitID() {
+        return unitID;
+    }
+
+    /**
+     * @param unitID the unitID to set
+     */
+    public void setUnitID(int unitID) {
+        this.unitID = unitID;
+    }
+
+    /**
+     * @return the name
+     */
+    public String getName() {
+        return name;
+    }
+
+    /**
+     * @param name the name to set
+     */
+    public void setName(String name) {
+        this.name = name;
+    }
+    
+    /**
+     * 
+     * @return 
+     */
+    public Set getUsers() {
+        return users;
+    }
+
+    /**
+     * 
+     * @param users 
+     */
+    public void setUsers(Set users) {
+        this.users = users;
+    }
+}
