+++ src/org/greenpole/entity/notification/NotificationWrapper.java
@@ -15,7 +15,10 @@ import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlSeeAlso;
 import javax.xml.bind.annotation.XmlTransient;
 import javax.xml.bind.annotation.XmlType;
-import org.greenpole.entity.model.ClientCompany;
+import org.greenpole.entity.model.clientcompany.ClientCompany;
+import org.greenpole.entity.model.user.Department;
+import org.greenpole.entity.model.user.Unit;
+import org.greenpole.entity.model.user.User;
 
 /**
  *
@@ -26,7 +29,7 @@ import org.greenpole.entity.model.ClientCompany;
 @XmlRootElement
 @XmlAccessorType(XmlAccessType.NONE)
 @XmlType(propOrder = {"code","messageTag","from","to","description","model"})
-@XmlSeeAlso(ClientCompany.class)
+@XmlSeeAlso({ClientCompany.class, User.class, Department.class, Unit.class})
 public class NotificationWrapper implements Serializable {
     @XmlElement
     private String code;
